refactor: set up a dispatch base hook to use in config set-up
