+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2023-11-27T08:09:00.537Z\n"
-"PO-Revision-Date: 2023-11-27T08:09:00.537Z\n"
+"POT-Creation-Date: 2023-12-13T11:14:15.701Z\n"
+"PO-Revision-Date: 2023-12-13T11:14:15.701Z\n"
 
 msgid "Select period(s)"
 msgstr "Select period(s)"
@@ -94,9 +94,6 @@ msgstr "Cannot be calculated"
 msgid "Exit configurations"
 msgstr "Exit configurations"
 
-msgid "Please provide a value"
-msgstr "Please provide a value"
-
 msgid "WHO Data Quality Annual Report"
 msgstr "WHO Data Quality Annual Report"
 
@@ -124,15 +121,6 @@ msgstr ""
 
 msgid "This configuration is used as the basis for the Annual Report."
 msgstr "This configuration is used as the basis for the Annual Report."
-
-msgid "Select period(s)"
-msgstr "Select period(s)"
-
-msgid "Cancel"
-msgstr "Cancel"
-
-msgid "Save"
-msgstr "Save"
 
 msgid ""
 "Group, Organisation unit, Organisation unit level, period, and comparison "
