+++ src/utils/configurations/configurationsContext.js
@@ -1,13 +1,9 @@
-import { useAlert, useDataEngine } from '@dhis2/app-runtime'
 import { Box, CenteredContent, NoticeBox } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useState, useContext } from 'react'
 import { LoadingSpinner } from '../../components/loading-spinner/LoadingSpinner.js'
-import { configurationsReducer } from './configurationsReducer.js'
-import {
-    useSetUpConfigurations,
-    DATASTORE_ID,
-} from './useSetUpConfigurations.js'
+import { useConfigurationsDispatchBase } from './useConfigurationsDispatchBase.js'
+import { useSetUpConfigurations } from './useSetUpConfigurations.js'
 
 const ErrorInfo = ({ errorMessage }) => (
     <Box height={'100%'}>
@@ -103,75 +99,23 @@ export const useRefetchConfigurations = () => {
     return refetch
 }
 
-const UPDATE_CONFIGURATIONS_MUTATION = {
-    resource: 'dataStore',
-    id: DATASTORE_ID,
-    type: 'update',
-    data: ({ newConfigurations }) => newConfigurations,
-}
-
 /**
  * Returns a `dispatch` function, intended to a `data` parameter with the shape
  * { type, payload }.
- * type is an action type, which can be found in configurationsReducer.js
- * payload is the required information to complete the relevant action
+ * type is an action type, which can be found in configurationsReducer.js;
+ * payload is the required information to complete the relevant action.
  *
- * Invokes the reducer function found in configurationsReducer.js to get a new
- * configurations object, then sets the local state to that new object,
- * attempts to sync that object with the server, and handles any errors
- *
- * Doesn't exactly use `useReducer` under the hood, but matches the same
- * behavior, pattern, and function signature, so the guidelines at
- * https://react.dev/reference/react/useReducer are relevant and helpful.
- * (`useState` is used here to help integrate with optimistic updates and
- * error checking)
+ * See useConfigurationsDispatchBase.js for more details
  */
 export const useConfigurationsDispatch = () => {
     const setConfigurations = useContext(SetConfigurationsContext)
-    const engine = useDataEngine()
-    const { show } = useAlert(
-        ({ errorMessage }) => 'Configurations update failed: ' + errorMessage,
-        { critical: true }
-    )
+    const dispatch = useConfigurationsDispatchBase(setConfigurations)
 
     if (!setConfigurations) {
         throw new Error(
             'useUpdateConfigurations must be used inside of a ConfigurationsProvider'
         )
     }
-
-    const dispatch = React.useCallback(
-        async ({ type, payload }) => {
-            let configurationsBackup
-            let newConfigurations
-            // set local configurations object (optimistically)
-            // use a function as an arg to get previous value without needing
-            // a dependency on `configurations`
-            setConfigurations((prevConfigurations) => {
-                // save a backup in case the mutation fails
-                configurationsBackup = prevConfigurations
-                // use reducer to get the new configurations object
-                newConfigurations = configurationsReducer(prevConfigurations, {
-                    type,
-                    payload,
-                })
-                return newConfigurations
-            })
-            try {
-                // update the configurations on the server
-                await engine.mutate(UPDATE_CONFIGURATIONS_MUTATION, {
-                    variables: { newConfigurations },
-                })
-            } catch (err) {
-                // if it fails, roll back to previous configurations locally
-                setConfigurations(configurationsBackup)
-                // and alert the error
-                show({ errorMessage: err.details?.message || err.message })
-                console.error(err, { details: err.details })
-            }
-        },
-        [setConfigurations, engine, show]
-    )
 
     return dispatch
 }
