+++ src/utils/configurations/index.js
@@ -1,3 +1,4 @@
 export * from './configurationsContext.js'
 export * from './configurationsReducer.js'
 export { checkNumeratorMappings } from './checkNumeratorMappings.js'
+export { DATASTORE_ID } from './constants.js'
